Renamed methods to use graph optimisation and a general tidy of the source files.
